mtd: spi: imx_m25pxx: use u8 for the id
This is done as a preparation for the refactor of the imx spi flash
drivers. As other two drivers have the idcode1 field of the struct
imx_spi_flash_params as u8.

+++ drivers/mtd/spi/imx_spi_nor_m25pxx.c
@@ -39,7 +39,7 @@ static u8 g_rx_buf[256];
 #define	SPI_FIFOSIZE		24
 
 struct imx_spi_flash_params {
-	u32		idcode1;
+	u8		idcode1;
 	u32		block_size;
 	u32		block_count;
 	u32		device_size;
